Revert "Rework last updated"
